Standardizes the use of receipt_url, error_url, and cancel_url
These properties were also called receipt_page_url, error_page_url, amd
cancel_page_url in some places.

Moves the property declarations into the base BasePaymentProcessor class.
