+++ ecommerce/extensions/payment/tests/processors/test_cybersource.py
@@ -58,9 +58,9 @@ class CybersourceTests(CybersourceMixin, PaymentProcessorTestCaseMixin, TestCase
             'amount': unicode(self.basket.total_incl_tax),
             'currency': self.basket.currency,
             'consumer_id': self.basket.owner.username,
-            'override_custom_receipt_page': '{}?orderNum={}'.format(self.processor.receipt_page_url,
+            'override_custom_receipt_page': '{}?orderNum={}'.format(self.processor.receipt_url,
                                                                     self.basket.order_number),
-            'override_custom_cancel_page': self.processor.cancel_page_url,
+            'override_custom_cancel_page': self.processor.cancel_url,
             'merchant_defined_data1': self.course.id,
             'merchant_defined_data2': self.CERTIFICATE_TYPE,
         }
