+++ ecommerce/settings/local.py
@@ -92,6 +92,7 @@ PAYMENT_PROCESSOR_CONFIG = {
             'payment_page_url': 'https://testsecureacceptance.cybersource.com/pay',
             'receipt_path': PAYMENT_PROCESSOR_RECEIPT_PATH,
             'cancel_path': PAYMENT_PROCESSOR_CANCEL_PATH,
+            'error_path': PAYMENT_PROCESSOR_ERROR_PATH,
             'send_level_2_3_details': True,
         },
         'paypal': {
