+++ ecommerce/templates/customer/emails/commtype_credit_receipt_body.html
@@ -39,7 +39,7 @@
                         <tr>
                             <td align="center" valign="middle"
                                 style="font-size:13px; background-color: #127eb1; border-radius: 5px; box-shadow: 0 2px #002D40;">
-                                <a href="{{ receipt_page_url }}" style="color: #ffffff; text-decoration: none;">
+                                <a href="{{ receipt_url }}" style="color: #ffffff; text-decoration: none;">
                                     {% trans "View Payment Information" %}
                                 </a>
                             </td>
